requirements.txt
pendulum<3
dbt-postgres==1.5.1
psycopg2-binary>=2.9
pandas
requests
python-dotenv
kaggle
sqlalchemy>=1.4
marshmallow-sqlalchemy>=0.28.0
=======
dbt-postgres==1.5.1
apache-airflow==2.7.1
apache-airflow-providers-postgres==5.4.0
psycopg2-binary>=2.9,<3.0
pandas
requests
python-dotenv
